ENH: Changing back since it looks like travis does not provide the environment variables to PR request runs

+++ act/tests/test_discovery.py
@@ -48,17 +48,18 @@ def test_get_armfile():
     uname = os.getenv('ARM_USERNAME')
     token = os.getenv('ARM_PASSWORD')
 
-    datastream = 'sgpmetE13.b1'
-    startdate = '2020-01-01'
-    enddate = startdate
-    outdir = os.getcwd() + '/data/'
+    if uname is not None:
+        datastream = 'sgpmetE13.b1'
+        startdate = '2020-01-01'
+        enddate = startdate
+        outdir = os.getcwd() + '/data/'
 
-    act.discovery.get_armfiles.download_data(uname, token, datastream,
-                                             startdate, enddate,
-                                             output=outdir)
-    files = glob.glob(outdir + datastream + '*20200101*cdf')
-    assert files is not None
-    assert 'sgpmetE13' in files[0]
+        act.discovery.get_armfiles.download_data(uname, token, datastream,
+                                                 startdate, enddate,
+                                                 output=outdir)
+        files = glob.glob(outdir + datastream + '*20200101*cdf')
+        assert files is not None
+        assert 'sgpmetE13' in files[0]
 
-    if files is not None:
-        os.remove(files[0])
+        if files is not None:
+            os.remove(files[0])
